backend_ml/requirements.txt
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.2.0
fastapi>=0.95.0
uvicorn>=0.22.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
